requirements.txt
Pillow==5.4.1
boto3==1.9.101
Flask==1.0.2
